Move to /receive-message dir
