+++ app/routes/receive-message.js
@@ -16,9 +16,9 @@ const inboundMessageMiddleware = require('../../lib/middleware/receive-message/m
 const getBotReplyMiddleware = require('../../lib/middleware/receive-message/reply-brain');
 const brainTemplateMiddleware = require('../../lib/middleware/receive-message/template-brain');
 const noReplyMiddleware = require('../../lib/middleware/receive-message/template-paused');
-const campaignMenuMiddleware = require('../../lib/middleware/campaign-menu');
-const campaignKeywordMiddleware = require('../../lib/middleware/campaign-keyword');
-const currentCampaignMiddleware = require('../../lib/middleware/campaign-current');
+const campaignMenuMiddleware = require('../../lib/middleware/receive-message/campaign-menu');
+const campaignKeywordMiddleware = require('../../lib/middleware/receive-message/campaign-keyword');
+const currentCampaignMiddleware = require('../../lib/middleware/receive-message/campaign-current');
 const parseAskSignupMiddleware = require('../../lib/middleware/receive-message/parse-ask-signup-response');
 const parseAskContinueMiddleware = require('../../lib/middleware/receive-message/parse-ask-continue-response');
 const askContinueMiddleware = require('../../lib/middleware/receive-message/template-ask-continue');
