+++ lib/middleware/receive-message/campaign-current.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const logger = require('heroku-logger');
-const Campaigns = require('../../app/models/Campaign');
+const Campaigns = require('../../../app/models/Campaign');
 
 module.exports = function getCurrentCampaign() {
   return (req, res, next) => {
