+++ lib/middleware/receive-message/campaign-keyword.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const Campaigns = require('../../app/models/Campaign');
+const Campaigns = require('../../../app/models/Campaign');
 
 module.exports = function getCampaignForKeyword() {
   return (req, res, next) => {
