+++ lib/middleware/receive-message/campaign-menu.js
@@ -1,7 +1,7 @@
 'use strict';
 
-const Campaigns = require('../../app/models/Campaign');
-const helpers = require('../helpers.js');
+const Campaigns = require('../../../app/models/Campaign');
+const helpers = require('../../helpers.js');
 
 module.exports = function getCampaignMenu() {
   return (req, res, next) => {
